% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{Simp}
\alias{Simp}
\title{Simulated Chim
 (Simp) Data}
\format{
\describe{
\item{Time}{times of observation} 
\item{X, Y}{coordinates}}
}
\source{
simulated data
}
\usage{
data(Simp)
}
\description{
This is a simulated movement track from a continuous auto-correlated process 
of total duration 60 time units with four behavioral phases that switch at 
times T=10, 20 and 50 from, variously, higher or lower velocity and longer or 
shorter characteristic time scale of autocorrelation. The original data was 
simulated with time intervals of 0.01 (leading to a complete track with 6000 
data points). 200 points were randomly sampled from the "true" movement 
track, such that the intervals between locations are random with mean 
interval 0.3 units.
}
\examples{
data(Simp)
plot(Simp)
}
\keyword{data}
